git-svn-id: https://korpling.german.hu-berlin.de/svn/saltnpepper/PepperModules/PAULAModules/trunk@6691 47085195-c50e-4317-8f43-7cb808db0f00

+++ src/main/java/de/hu_berlin/german/korpling/saltnpepper/pepperModules/paula/Salt2PAULAMapper.java
@@ -179,7 +179,8 @@ public class Salt2PAULAMapper implements PAULAXMLStructure, FilenameFilter
 		if (resourcePath != null)
 		{
 			File DTDDirectory = new File(resourcePath.toFileString()+File.separator+"dtd_09");
-			if (DTDDirectory.listFiles(this)!= null)
+			if (	(DTDDirectory.exists())&&
+					(DTDDirectory.listFiles(this)!= null))
 			{
 				for (File DTDFile : DTDDirectory.listFiles(this)){
 					copyFile(URI.createFileURI(DTDFile.getAbsolutePath()), documentPath.toFileString());
